# src/backend/src/Controllers/TemperatureApi.cs
/*
 * SmartHomeApi
 *
 * Documentation of all Api-Calls the SmartHome Backend can handle.
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class TemperatureApiController : ControllerBase
    { 
        /// <summary>
        /// Gets the last emmited Temperature.
        /// </summary>
        
        /// <response code="200">OK</response>
        [HttpGet]
        [Route("/v1/temperature/getCurrentTemperature")]
        [ValidateModelState]
        [SwaggerOperation("GetCurrentTemperature")]
        [SwaggerResponse(statusCode: 200, type: typeof(TemperatureData), description: "OK")]
        public virtual IActionResult GetCurrentTemperature()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(TemperatureData));

            string exampleJson = null;
            exampleJson = "{\n  \"valueF\" : 6.02745618307040320615897144307382404804229736328125,\n  \"timeStamp\" : \"2000-01-23T04:56:07.000+00:00\",\n  \"valueC\" : 0.80082819046101150206595775671303272247314453125\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<TemperatureData>(exampleJson)
            : default(TemperatureData);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Emmit a change in the current temperature
        /// </summary>
        
        /// <param name="body">Pet object that needs to be added to the store</param>
        /// <response code="200">OK</response>
        [HttpPut]
        [Route("/v1/temperature/setCurrentTemperature")]
        [ValidateModelState]
        [SwaggerOperation("SetCurrentTemperature")]
        public virtual IActionResult SetCurrentTemperature([FromBody]TemperatureData body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);


            throw new NotImplementedException();
        }
    }
}