# src/backend/src/Models/TemperatureData.cs
/*
 * SmartHomeApi
 *
 * Documentation of all Api-Calls the SmartHome Backend can handle.
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class TemperatureData : IEquatable<TemperatureData>
    { 
        /// <summary>
        /// Gets or Sets ValueC
        /// </summary>
        [DataMember(Name="valueC")]
        public decimal? ValueC { get; set; }

        /// <summary>
        /// Gets or Sets ValueF
        /// </summary>
        [DataMember(Name="valueF")]
        public decimal? ValueF { get; set; }

        /// <summary>
        /// Gets or Sets TimeStamp
        /// </summary>
        [DataMember(Name="timeStamp")]
        public DateTime? TimeStamp { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class TemperatureData {\n");
            sb.Append("  ValueC: ").Append(ValueC).Append("\n");
            sb.Append("  ValueF: ").Append(ValueF).Append("\n");
            sb.Append("  TimeStamp: ").Append(TimeStamp).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((TemperatureData)obj);
        }

        /// <summary>
        /// Returns true if TemperatureData instances are equal
        /// </summary>
        /// <param name="other">Instance of TemperatureData to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(TemperatureData other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    ValueC == other.ValueC ||
                    ValueC != null &&
                    ValueC.Equals(other.ValueC)
                ) && 
                (
                    ValueF == other.ValueF ||
                    ValueF != null &&
                    ValueF.Equals(other.ValueF)
                ) && 
                (
                    TimeStamp == other.TimeStamp ||
                    TimeStamp != null &&
                    TimeStamp.Equals(other.TimeStamp)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (ValueC != null)
                    hashCode = hashCode * 59 + ValueC.GetHashCode();
                    if (ValueF != null)
                    hashCode = hashCode * 59 + ValueF.GetHashCode();
                    if (TimeStamp != null)
                    hashCode = hashCode * 59 + TimeStamp.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(TemperatureData left, TemperatureData right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(TemperatureData left, TemperatureData right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}